+++ CHANGE_LOG.md
@@ -49,3 +49,6 @@ All notable changes to this project will be documented in this file.## [Unreleas
 - Added a fix for auth chat to connect to the same chat after refresh.
 - Fixed the sound notification when the footer is hidden.
 - Replaced proactive chat pane close button with header close button.
+- Adding `localStorage` support for widget
+- Adding `typescript` sample
+- Adding support for customizing anchor tag color in webchat